+++ Tp_Ecom/src/fr/adaming/service/IClientService.java
@@ -15,7 +15,7 @@ public interface IClientService {
 
 	public Client updateClient(Client c);
 
-	public int deleteClient(int id);
+	public int deleteClient(Long id);
 
-	public Client getClientById(int id);
+	public Client getClientById(Long id);
 }
