+++ Tp_Ecom/src/fr/adaming/service/ICommandeService.java
@@ -15,7 +15,7 @@ public interface ICommandeService {
 	
 	public Commande updateCommande(Commande c);
 	
-	public int deleteCommande(int id);
+	public int deleteCommande(Long id);
 
-	public Commande getCommandeById(int id);
+	public Commande getCommandeById(Long id);
 }
